LoopTree Workload API: Tensor Size in Floating Point (#305)
* Add functions to compute equivalent ranks and tensor size

* Remove boost dependency from isl-functions

* Add utility method to calculate operation space volume

* Fix parsing bug where tile_size is looked up instead of tile_shape

* Remove dependency on boost log and filesystem libraries

* LoopTree to Python interface update

* [looptree/dependency-analyzer] WIP pairwise dependency analyzer

* [looptree] Analyze shared input reuse rank equivalence

* [looptree] Bugfix for dependency analyzer

* [looptree] Update GetTensorSize and GetOspaceVolume to return double

---------

+++ include/workload/fused-workload.hpp
@@ -77,9 +77,9 @@ class FusedWorkload
 
   std::tuple<int, int> GetRankShape(DimensionId einsum_rank) const;
 
-  int GetTensorSize(DataSpaceId dspace) const;
+  double GetTensorSize(DataSpaceId dspace) const;
 
-  int GetOspaceVolume(EinsumId einsum) const;
+  double GetOspaceVolume(EinsumId einsum) const;
 
  private:
   std::map<std::string, EinsumId> einsum_name_to_id_;
